+++ changelog.md
@@ -1,5 +1,6 @@
 # 0.128
 
+- Adding decoy makers
 - Take 000 out of regular flow on new game screen/highlight as tutorial; clean up new game page slightly
 - Bug fix: assault move should also exhaust carried weapons
 - Updated/Promoted 406